% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{apply_custom_link_priors}
\alias{apply_custom_link_priors}
\title{Apply custom prior specifications for link parameters}
\usage{
apply_custom_link_priors(result, link_prior)
}
\arguments{
\item{result}{Current result list}

\item{link_prior}{List of custom prior specifications}
}
\value{
Updated result list with custom priors
}
\description{
Apply custom prior specifications for link parameters
}
\keyword{internal}
