% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_cdfs.R
\name{aranda_ordaz_cdf}
\alias{aranda_ordaz_cdf}
\title{Aranda-Ordaz asymmetric link CDF}
\usage{
aranda_ordaz_cdf(x, lambda)
}
\arguments{
\item{x}{Numeric vector}

\item{lambda}{Shape parameter (lambda > 0)}
}
\value{
CDF values
}
\description{
F(x; lambda) = 1 - (1 + exp(x))^(-lambda)
}
\details{
Special cases:
\itemize{
\item lambda = 1: logit
\item lambda -> 0: cloglog
}

Reference: Aranda-Ordaz (1981) Biometrika
}
\keyword{internal}
