% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_cdfs.R
\name{clm_cdf}
\alias{clm_cdf}
\title{Unified CDF dispatcher for clmstan}
\usage{
clm_cdf(x, link, link_param = list())
}
\arguments{
\item{x}{Numeric vector of values}

\item{link}{Link function name (character)}

\item{link_param}{List of link parameters (for flexible links)}
}
\value{
CDF values F(x)
}
\description{
Computes the cumulative distribution function (CDF) for any supported
link function. This is the R equivalent of \code{unified_F()} in Stan.
}
\keyword{internal}
