% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmstan-class.R
\name{clmstan-class}
\alias{clmstan-class}
\title{clmstan S3 Class}
\description{
The \code{clmstan} class represents a fitted cumulative link model.
It contains the CmdStanR fit object and additional metadata.
}
\section{Slots}{

\describe{
\item{fit}{The CmdStanMCMC object from cmdstanr}
\item{formula}{The model formula}
\item{data}{The original data frame}
\item{link}{The link function used}
\item{base}{The base distribution (for SP link)}
\item{threshold}{The threshold structure}
\item{link_param}{Link parameter settings (for flexible links)}
\item{full}{TRUE if link parameters were estimated (Bayesian inference),
FALSE if they were fixed at user-specified values}
\item{K}{Number of response categories (cached from data)}
\item{N}{Number of observations (extracted from data for efficiency)}
\item{P}{Number of predictors (extracted from design matrix)}
}
}

\concept{clmstan methods}
