% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_cdfs.R
\name{cloglog_cdf}
\alias{cloglog_cdf}
\title{Complementary log-log CDF}
\usage{
cloglog_cdf(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
CDF values
}
\description{
F(x) = 1 - exp(-exp(x))
}
\details{
This corresponds to the Gumbel (maximum) distribution.
}
\keyword{internal}
