% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{compute_probs_all}
\alias{compute_probs_all}
\title{Compute category probabilities for all draws}
\usage{
compute_probs_all(draws, X, link, base_link_param, K, full = FALSE)
}
\arguments{
\item{draws}{Output from extract_prediction_draws()}

\item{X}{Design matrix (N x P)}

\item{link}{Link function name}

\item{base_link_param}{Fixed link parameters}

\item{K}{Number of categories}

\item{full}{Whether link parameters were estimated}
}
\value{
Array of S x N x K probabilities
}
\description{
Compute category probabilities for all draws
}
\keyword{internal}
