% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{compute_probs_single}
\alias{compute_probs_single}
\title{Compute category probabilities for a single draw}
\usage{
compute_probs_single(c_vec, eta_vec, link, link_param, K)
}
\arguments{
\item{c_vec}{Vector of K-1 cutpoints}

\item{eta_vec}{Vector of N linear predictors}

\item{link}{Link function name}

\item{link_param}{Link parameters (for this draw)}

\item{K}{Number of categories}
}
\value{
Matrix of N x K probabilities
}
\description{
Compute category probabilities for a single draw
}
\keyword{internal}
