% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_waic.R
\name{compute_r_eff}
\alias{compute_r_eff}
\title{Compute relative effective sample size for log-likelihood}
\usage{
compute_r_eff(object, log_lik, cores = 1)
}
\arguments{
\item{object}{A clmstan object}

\item{log_lik}{Log-likelihood matrix (S x N)}

\item{cores}{Number of cores for parallel computation}
}
\value{
A vector of relative effective sample sizes (length N)
}
\description{
Compute relative effective sample size for log-likelihood
}
\keyword{internal}
