% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_link_param_draws}
\alias{extract_link_param_draws}
\title{Extract link parameter draws for full model}
\usage{
extract_link_param_draws(object)
}
\arguments{
\item{object}{A clmstan object}
}
\value{
List of link parameter vectors
}
\description{
Extract link parameter draws for full model
}
\keyword{internal}
