% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_link_params_summary}
\alias{extract_link_params_summary}
\title{Extract posterior summary for link parameters}
\usage{
extract_link_params_summary(object, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{object}{A clmstan object}

\item{probs}{Quantile probabilities}
}
\value{
A data frame with posterior summaries, or NULL if no estimated params
}
\description{
Extract posterior summary for link parameters
}
\keyword{internal}
