% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_waic.R
\name{extract_log_lik}
\alias{extract_log_lik}
\title{Extract log-likelihood matrix from clmstan object}
\usage{
extract_log_lik(object)
}
\arguments{
\item{object}{A clmstan object}
}
\value{
An S x N matrix of log-likelihood values, where S is the number
of posterior samples and N is the number of observations.
}
\description{
Extract log-likelihood matrix from clmstan object
}
\keyword{internal}
