% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_prediction_draws}
\alias{extract_prediction_draws}
\title{Extract posterior draws needed for prediction}
\usage{
extract_prediction_draws(object, ndraws = NULL)
}
\arguments{
\item{object}{A clmstan object}

\item{ndraws}{Number of draws to extract}
}
\value{
List with c (cutpoints), beta (coefficients), link_params (if full model)
}
\description{
Extract posterior draws needed for prediction
}
\keyword{internal}
