% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_y_rep}
\alias{extract_y_rep}
\title{Extract y_rep from Stan fit}
\usage{
extract_y_rep(object, ndraws = NULL)
}
\arguments{
\item{object}{A clmstan object}

\item{ndraws}{Number of draws to extract}
}
\value{
S x N integer matrix
}
\description{
Extract y_rep from Stan fit
}
\keyword{internal}
