% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{format_dist}
\alias{format_dist}
\title{Format a distribution object as a string}
\usage{
format_dist(x)
}
\arguments{
\item{x}{A clm_dist object}
}
\value{
A character string representation
}
\description{
Format a distribution object as a string
}
\keyword{internal}
