% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_base_link_params}
\alias{get_base_link_params}
\title{Get base (fixed) link parameters from object}
\usage{
get_base_link_params(object)
}
\arguments{
\item{object}{A clmstan object}
}
\value{
List of fixed link parameters
}
\description{
Get base (fixed) link parameters from object
}
\keyword{internal}
