% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_coef_names}
\alias{get_coef_names}
\title{Get coefficient names from design matrix}
\usage{
get_coef_names(object)
}
\arguments{
\item{object}{A clmstan object}
}
\value{
A character vector of coefficient names
}
\description{
Get coefficient names from design matrix
}
\keyword{internal}
