% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{get_default_link_prior}
\alias{get_default_link_prior}
\title{Get default prior for a link parameter}
\usage{
get_default_link_prior(link, param)
}
\arguments{
\item{link}{A character string specifying the link function}

\item{param}{A character string specifying the parameter name}
}
\value{
A list with prior specification (family, args)
}
\description{
Get default prior for a link parameter
}
\keyword{internal}
