% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_link_param_for_draw}
\alias{get_link_param_for_draw}
\title{Get link parameters for a specific draw}
\usage{
get_link_param_for_draw(link_param_draws, s, base_param)
}
\arguments{
\item{link_param_draws}{List of link parameter vectors}

\item{s}{Draw index}

\item{base_param}{Base (fixed) parameters}
}
\value{
Link parameters for draw s
}
\description{
Get link parameters for a specific draw
}
\keyword{internal}
