% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{get_link_params}
\alias{get_link_params}
\title{Get required parameters for a link function}
\usage{
get_link_params(link)
}
\arguments{
\item{link}{A character string specifying the link function}
}
\value{
A character vector of required parameter names
}
\description{
Get required parameters for a link function
}
\keyword{internal}
