% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{get_link_type}
\alias{get_link_type}
\title{Get link type number for Stan}
\usage{
get_link_type(link)
}
\arguments{
\item{link}{A character string specifying the link function}
}
\value{
An integer (1-11) representing the link type
}
\description{
Converts a link function name to the integer code used in Stan.
}
\details{
Link type mapping:
\itemize{
\item 1: logit
\item 2: probit
\item 3: cloglog
\item 4: loglog
\item 5: cauchit
\item 6: tlink
\item 7: aranda_ordaz
\item 8: sp
\item 9: log_gamma
\item 10: gev
\item 11: aep
}
}
\keyword{internal}
