% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{get_model_name}
\alias{get_model_name}
\title{Get Stan model name based on threshold and full model flag}
\usage{
get_model_name(threshold, full = FALSE)
}
\arguments{
\item{threshold}{A character string specifying the threshold structure}

\item{full}{Logical indicating whether link parameters are estimated}
}
\value{
A character string with the Stan model name
}
\description{
Get Stan model name based on threshold and full model flag
}
\keyword{internal}
