% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_threshold_names}
\alias{get_threshold_names}
\title{Generate threshold names in ordinal::clm style}
\usage{
get_threshold_names(object)
}
\arguments{
\item{object}{A clmstan object}
}
\value{
A character vector of threshold names
}
\description{
Creates threshold names like "1|2", "2|3", etc. based on factor levels.
}
\keyword{internal}
