% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmstan-class.R
\name{is.clmstan}
\alias{is.clmstan}
\title{Check if object is clmstan}
\usage{
is.clmstan(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if x is a clmstan object
}
\description{
Check if object is clmstan
}
