% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{is_flexible_link}
\alias{is_flexible_link}
\title{Check if a link function requires parameters}
\usage{
is_flexible_link(link)
}
\arguments{
\item{link}{A character string specifying the link function}
}
\value{
TRUE if the link requires additional parameters
}
\description{
Check if a link function requires parameters
}
\keyword{internal}
