% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_cdfs.R
\name{loglog_cdf}
\alias{loglog_cdf}
\title{Log-log CDF}
\usage{
loglog_cdf(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
CDF values
}
\description{
F(x) = exp(-exp(-x))
}
\details{
This corresponds to the Gumbel (minimum) distribution.
It is the reflection of cloglog: loglog(x) = 1 - cloglog(-x)
}
\keyword{internal}
