% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Create design matrix}
\usage{
make_design_matrix(formula, data)
}
\arguments{
\item{formula}{A formula object or terms (RHS only)}

\item{data}{A data frame}
}
\value{
A design matrix (N x P)
}
\description{
Creates the design matrix for predictors (without intercept column).
}
\keyword{internal}
