% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{map_class_to_params}
\alias{map_class_to_params}
\title{Map Parameter Class and Distribution to Legacy Parameter Names}
\usage{
map_class_to_params(class, dist)
}
\arguments{
\item{class}{The parameter class}

\item{dist}{A clm_dist object}
}
\value{
A named list of legacy parameter values
}
\description{
Map Parameter Class and Distribution to Legacy Parameter Names
}
\keyword{internal}
