% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{needs_full_model}
\alias{needs_full_model}
\title{Check if any link parameter requires estimation}
\usage{
needs_full_model(link_param)
}
\arguments{
\item{link_param}{A list of link parameters}
}
\value{
TRUE if any parameter is set to "estimate"
}
\description{
Check if any link parameter requires estimation
}
\keyword{internal}
