% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmstan-class.R
\name{new_clmstan}
\alias{new_clmstan}
\title{Create a clmstan object (internal constructor)}
\usage{
new_clmstan(
  fit,
  formula,
  data,
  link,
  base,
  threshold,
  link_param = NULL,
  full = FALSE,
  K,
  N,
  P
)
}
\arguments{
\item{fit}{A CmdStanMCMC object}

\item{formula}{The model formula}

\item{data}{The original data}

\item{link}{The link function}

\item{base}{The base distribution (for SP link)}

\item{threshold}{The threshold structure}

\item{link_param}{Link parameter settings}

\item{full}{Whether full model was used}

\item{K}{Number of categories}

\item{N}{Number of observations}

\item{P}{Number of predictors}
}
\value{
An object of class "clmstan"
}
\description{
This is an internal constructor called by clm_stan(). Users should not
call this function directly.
}
\keyword{internal}
