% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{parse_clm_formula}
\alias{parse_clm_formula}
\title{Parse formula for CLM}
\usage{
parse_clm_formula(formula, data)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}
}
\value{
A list with response and predictor_formula
}
\description{
Parse formula for CLM
}
\keyword{internal}
