% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_link_params}
\alias{prepare_link_params}
\title{Prepare link parameters for Stan}
\usage{
prepare_link_params(link, link_param = NULL)
}
\arguments{
\item{link}{Link function name}

\item{link_param}{A list of user-specified link parameters}
}
\value{
A list with all link parameters (df, lambda, xi, r, base_type, theta1, theta2)
}
\description{
Fills in default values for link parameters and validates them.
}
\keyword{internal}
