% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_link_params_full}
\alias{prepare_link_params_full}
\title{Prepare link parameters for full Stan model}
\usage{
prepare_link_params_full(link, link_param = NULL, link_prior = NULL)
}
\arguments{
\item{link}{Link function name}

\item{link_param}{A list of user-specified link parameters}

\item{link_prior}{A list of custom prior specifications}
}
\value{
A list with estimation flags, fixed values, and prior hyperparameters
}
\description{
Determines which parameters to estimate vs. fix, and sets up priors.
}
\keyword{internal}
