% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prepare_prediction_matrix}
\alias{prepare_prediction_matrix}
\title{Prepare design matrix for prediction}
\usage{
prepare_prediction_matrix(object, newdata)
}
\arguments{
\item{object}{A clmstan object}

\item{newdata}{Data frame for prediction}
}
\value{
Design matrix (N x P)
}
\description{
Prepare design matrix for prediction
}
\keyword{internal}
