% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.clm_dist}
\alias{print.clm_dist}
\title{Print method for clm_dist objects}
\usage{
\method{print}{clm_dist}(x, ...)
}
\arguments{
\item{x}{A clm_dist object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for clm_dist objects
}
