% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.clm_prior}
\alias{print.clm_prior}
\title{Print method for clm_prior objects}
\usage{
\method{print}{clm_prior}(x, ...)
}
\arguments{
\item{x}{A clm_prior object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input \code{clm_prior} object.
}
\description{
Print method for clm_prior objects
}
