% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.clm_prior_spec}
\alias{print.clm_prior_spec}
\title{Print method for clm_prior_spec objects}
\usage{
\method{print}{clm_prior_spec}(x, ...)
}
\arguments{
\item{x}{A clm_prior_spec object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for clm_prior_spec objects
}
