% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.clmstan}
\alias{print.clmstan}
\title{Print method for clmstan objects}
\usage{
\method{print}{clmstan}(x, ...)
}
\arguments{
\item{x}{A clmstan object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for clmstan objects
}
