% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.clmstan}
\alias{print.summary.clmstan}
\title{Print method for summary.clmstan objects}
\usage{
\method{print}{summary.clmstan}(x, ...)
}
\arguments{
\item{x}{A summary.clmstan object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for summary.clmstan objects
}
