% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summarize_class_draws}
\alias{summarize_class_draws}
\title{Summarize class prediction draws}
\usage{
summarize_class_draws(class_draws, robust = FALSE, probs = c(0.025, 0.975))
}
\arguments{
\item{class_draws}{S x N matrix of predicted classes}

\item{robust}{Use median instead of mean}

\item{probs}{Quantile probabilities}
}
\value{
Data frame with summary statistics
}
\description{
Summarize class prediction draws
}
\keyword{internal}
