% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summarize_probs_draws}
\alias{summarize_probs_draws}
\title{Summarize probability draws}
\usage{
summarize_probs_draws(probs_array, robust = FALSE)
}
\arguments{
\item{probs_array}{S x N x K array of probability draws}

\item{robust}{Use median instead of mean}
}
\value{
Data frame with mean probabilities per category
}
\description{
Summarize probability draws
}
\keyword{internal}
