% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{supported_links}
\alias{supported_links}
\title{Get supported link functions}
\usage{
supported_links(type = c("all", "standard", "flexible"))
}
\arguments{
\item{type}{Character string specifying which links to return:
\itemize{
\item \code{"all"} (default): All supported link functions
\item \code{"standard"}: Standard links without additional parameters
\item \code{"flexible"}: Flexible links with additional parameters
}}
}
\value{
A character vector of supported link function names
}
\description{
Get supported link functions
}
\examples{
supported_links()
supported_links("standard")
supported_links("flexible")
}
