% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{supported_thresholds}
\alias{supported_thresholds}
\title{Get supported threshold structures}
\usage{
supported_thresholds()
}
\value{
A character vector of supported threshold structure names
}
\description{
Get supported threshold structures
}
\examples{
supported_thresholds()
}
