% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_full_params}
\alias{validate_full_params}
\title{Validate full model parameters}
\usage{
validate_full_params(link, params)
}
\arguments{
\item{link}{Link function name}

\item{params}{Full model parameters list}
}
\description{
Checks that estimation flags make sense for the specified link.
}
\keyword{internal}
