% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{validate_link}
\alias{validate_link}
\title{Check if a link function is valid}
\usage{
validate_link(link)
}
\arguments{
\item{link}{A character string specifying the link function}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Check if a link function is valid
}
\keyword{internal}
