% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_link_params}
\alias{validate_link_params}
\title{Validate link parameters}
\usage{
validate_link_params(link, params)
}
\arguments{
\item{link}{Link function name}

\item{params}{A list of link parameters}
}
\description{
Validate link parameters
}
\keyword{internal}
