% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{validate_prediction_input}
\alias{validate_prediction_input}
\title{Validate prediction inputs}
\usage{
validate_prediction_input(object, newdata, ndraws)
}
\arguments{
\item{object}{A clmstan object}

\item{newdata}{New data frame (or NULL)}

\item{ndraws}{Number of draws to use}
}
\description{
Validate prediction inputs
}
\keyword{internal}
