% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{validate_prior}
\alias{validate_prior}
\title{Validate prior specification for a model}
\usage{
validate_prior(prior, threshold, link, link_param)
}
\arguments{
\item{prior}{A clm_prior object or NULL}

\item{threshold}{The threshold structure}

\item{link}{The link function}

\item{link_param}{The link parameters}
}
\value{
TRUE if valid, otherwise throws a warning
}
\description{
Checks that the prior specification is compatible with the model settings.
}
\keyword{internal}
