% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{validate_prior_class_dist}
\alias{validate_prior_class_dist}
\title{Validate that the distribution is compatible with the parameter class}
\usage{
validate_prior_class_dist(class, dist)
}
\arguments{
\item{class}{The parameter class}

\item{dist}{A clm_dist object}
}
\description{
Validate that the distribution is compatible with the parameter class
}
\keyword{internal}
