% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{validate_prior_values}
\alias{validate_prior_values}
\title{Validate prior values}
\usage{
validate_prior_values(prior)
}
\arguments{
\item{prior}{A list of prior values}
}
\description{
Validate prior values
}
\keyword{internal}
