% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.clm_prior_list}
\alias{print.clm_prior_list}
\title{Print method for clm_prior_list objects}
\usage{
\method{print}{clm_prior_list}(x, ...)
}
\arguments{
\item{x}{A clm_prior_list object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input \code{clm_prior_list} object.
}
\description{
Print method for clm_prior_list objects
}
